Make application a class instead of an interface, should be only 1 implementor
