+++ corona-ide-core-api/src/main/java/com/coronaide/core/Application.java
@@ -11,25 +11,80 @@
 package com.coronaide.core;
 
 import java.nio.file.Path;
+import java.util.Objects;
+
+import javax.annotation.Nullable;
+import javax.annotation.concurrent.Immutable;
+
+import com.coronaide.core.service.IApplicationService;
 
 /**
- * Represents the Corona IDE application instance currently running
+ * Represents the Corona IDE application instance currently running.
+ *
+ * <p>
+ * Clients are not intended to create instances of this class - retrieve the application instance from the
+ * {@link IApplicationService} instance
  *
  * @author romeara
  * @since 0.1
  */
-public interface Application {
+@Immutable
+public final class Application {
+
+    private static Version VERSION = new Version(0, 0, 1);
+
+    private Path workingDirectory;
+
+    /**
+     * @param workingDirectory
+     *            Directory all application generated/managed files should be placed within
+     * @since 0.1
+     */
+    public Application(Path workingDirectory) {
+        this.workingDirectory = Objects.requireNonNull(workingDirectory);
+    }
 
     /**
      * @return The current running version of the application
      * @since 0.1
      */
-    Version getVersion();
+    public Version getVersion() {
+        return VERSION;
+    }
 
     /**
      * @return Directory where all files managed by the Corona IDE application are located
      * @since 0.1
      */
-    Path getWorkingDirectory();
+    public Path getWorkingDirectory() {
+        return workingDirectory;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(getVersion(), getWorkingDirectory());
+    }
+
+    @Override
+    public boolean equals(@Nullable Object obj) {
+        boolean result = false;
+
+        if (obj instanceof Application) {
+            Application compare = (Application) obj;
+
+            result = Objects.equals(compare.getVersion(), getVersion())
+                    && Objects.equals(compare.getWorkingDirectory(), getWorkingDirectory());
+        }
+
+        return result;
+    }
+
+    @Override
+    public String toString() {
+        return new StringBuilder().append(getClass().getSimpleName()).append('{')
+                .append("version").append('=').append(getVersion()).append(',')
+                .append("workingDirectory").append('=').append(getWorkingDirectory())
+                .append('}').toString();
+    }
 
 }
